requirements.txt
torch
sentence_transformers
rouge_score
bert_score
google.generativeai
dotenv
huggingface_hub[hf_xet]
seaborn
